+++ lib/bh/core_ext/rails/form_for_helper.rb
@@ -10,7 +10,14 @@ module Bh
       def form_for(record, options = {}, &block)
         options[:layout] ||= 'navbar' if Bh::Stack.find(Bh::Navbar)
         add_form_options!(options) if options[:layout]
-        super record, options, &block
+        html = super record, options, &block
+
+        if Bh::Stack.find(Bh::Nav)
+          container = Bh::Base.new(self) { html }
+          container.render_tag :li
+        else
+          html
+        end
       end
 
     private