+++ spec/rails/form_for_helper_spec.rb
@@ -7,18 +7,19 @@ describe 'form_for' do
   let(:protect_against_forgery?) { false }
   attr_accessor :output_buffer
   let(:form) { form_for(:object, options.merge(url: '/')) {} }
+  let(:options) { {} }
 
-  describe 'without a :layout option' do
-    let(:options) { {} }
+  specify 'by default, does not apply Bootstrap attributes to the form' do
+    expect(form).not_to include 'role="form"'
+  end
 
-    specify 'by default, does not apply Bootstrap attributes to the form' do
-      expect(form).not_to include 'role="form"'
-    end
+  specify 'wrapped in navbar, applies Bootstrap attributes of a navbar form' do
+    bh.navbar { expect(form).to include 'role="form"' }
+    bh.navbar { expect(form).to include 'class="navbar-form"' }
+  end
 
-    specify 'wrapped in navbar, applies Bootstrap attributes of a navbar form' do
-      bh.navbar { expect(form).to include 'role="form"' }
-      bh.navbar { expect(form).to include 'class="navbar-form"' }
-    end
+  specify 'wrapped in nav, surrounds the form in a <li> item' do
+    bh.nav { expect(form).to match %r{^<li><form.+?</form></li>$} }
   end
 
   describe 'with layout: :horizontal' do