+++ lib/bh/core_ext/rails/form_for_helper.rb
@@ -8,8 +8,16 @@ module Bh
       include ActionView::Helpers::FormHelper # for form_for
 
       def form_for(record, options = {}, &block)
+        options[:layout] ||= 'navbar' if Bh::Stack.find(Bh::Navbar)
         add_form_options!(options) if options[:layout]
-        super record, options, &block
+        html = super record, options, &block
+
+        if Bh::Stack.find(Bh::Nav)
+          container = Bh::Base.new(self) { html }
+          container.render_tag :li
+        else
+          html
+        end
       end
 
     private
@@ -23,6 +31,7 @@ module Bh
 
       def class_for(layout)
         case layout.to_s
+          when 'navbar' then 'navbar-form'
           when 'inline' then 'form-inline'
           when 'horizontal' then 'form-horizontal'
         end