+++ spec/shared/button_to_helper.rb
@@ -3,6 +3,7 @@ shared_examples_for 'the button_to helper' do
   all_tests_pass_with 'the :size button_to option'
   all_tests_pass_with 'the :layout button_to option'
   all_tests_pass_with 'the button wrapped in navbar'
+  all_tests_pass_with 'the button wrapped in nav'
 end
 
 #--
@@ -38,5 +39,13 @@ shared_examples_for 'the button wrapped in navbar' do
   specify 'adds the "navbar-form" class to the form' do
     html = %r{^<form.+class="navbar-form"}
     bh.navbar { expect(button_to: {context: :default}).to generate html }
+  end
+end
+
+
+shared_examples_for 'the button wrapped in nav' do
+  specify 'surrounds the form in a <li> item' do
+    html = %r{^<li><form.+?</form></li>}
+    bh.nav { expect(:button_to).to generate html }
   end
 end